split device into files
